+++ dejafu-tests/lib/Integration/SCT.hs
@@ -1,5 +1,3 @@
-{-# OPTIONS_GHC -Wno-deprecations #-}
-
 module Integration.SCT where
 
 import           Control.Concurrent.Classy hiding (check)
@@ -89,24 +87,29 @@ resultsSetTests = toTestList
     , testCase "Proper results from resultsSet'" $ do
         tested <- resultsSet' defaultWay defaultMemType testAction
         results @=? tested
-    , testCase "Proper results from resultsSetDiscard" $ do
+    , testCase "Proper results from resultsWithSettings" $ do
         tested <-
-          resultsSetDiscard discarder defaultWay defaultMemType testAction
+          resultsSetWithSettings rsSettings testAction
         resultsWithDiscard @=? tested
-    , testCase "Proper results from resultsSetDiscard'" $ do
+    , testCase "Proper results from resultsWithSettings'" $ do
         tested <-
-          resultsSetDiscard' discarder defaultWay defaultMemType testAction
+          resultsSetWithSettings' rsSettings testAction
         resultsWithDiscard @=? tested
     ]
   where
     results = S.fromList $ map Right [1, 2] ++ [Left Deadlock]
     resultsWithDiscard = S.fromList [Right 2, Left Deadlock]
-    discarder (Right 1) = Just DiscardResultAndTrace
-    discarder (Right 2) = Just DiscardTrace
-    discarder _ = Nothing
+
     testAction = do
       mvar <- newEmptyMVarInt
       _ <- fork $ putMVar mvar 1
       _ <- fork $ putMVar mvar 2
       _ <- fork $ mapM_ (\_ -> void $ takeMVar mvar) [1 :: Int, 2]
       readMVar mvar
+
+    rsSettings =
+      let d (Right 1) = Just DiscardResultAndTrace
+          d (Right 2) = Just DiscardTrace
+          d _ = Nothing
+      in set ldiscard (Just d) $
+         fromWayAndMemType defaultWay defaultMemType